requirements.txt
Flask==3.0.3
gunicorn==22.0.0
gspread==6.1.2
google-auth==2.33.0
requests==2.32.3
openai==1.40.0
